+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
 
 * Features
+  * Paperclip registry inspection
   * ActiveRecord schema cache and schema_cache.dump inspection
   * Configuration control of notification_tracking